+++ week3/BusinessCard/app/src/main/java/com/manuflowers/businesscard/utils/Enums.kt
@@ -1,4 +1,4 @@
-package com.manuflowers.businesscard
+package com.manuflowers.businesscard.utils
 
 enum class ThemeMode {
     LIGHT, DARK, SYSTEM, BATTERY
